Browse path related changes.
1. Replaced numeric nodeIds with browse names.
Exception: Browse node given by the application will be shown in
the given format. It means, if application wants to browse two nodes
(Ex: N1 and N2), it currently gives the node id of both nodes to
browseNode(). NodeId can be either numeric or string.
If nodeId is numeric, the corresponding path in browse results will
start with the numeric value of the nodeid following by the browse names
for the rest of the nodes. Ex: /84/Types/ReferenceTypes

2. Pass browse path to application.
The complete browse path including the browse name will be passed in the
browse response callback. [EdgeVersatility in EdgeResponse will have the
unsigned char format string for the browse path.]

Signed-off-by: Senthil Kumar G S <senthil.gs@samsung.com>